feat(test): 重整hystrix目录测试文件结构
将2个测试文件重组为1个标准命名的测试文件：
- 删除: performance_test.go (性能测试文件)
- 合并到: hystrix_test.go (熔断器模式测试)
- 保持34个测试函数和8个基准测试
- 涵盖熔断器状态管理、故障检测、回退机制、性能优化功能
